requirements.txt
pycryptodome>=3.15.0
frida-tools>=10.0.0